[NETCFGX] Move the NetClassInstaller to a separate file.
